requirements.txt
Flask==2.2.2
python-dotenv==0.21.0
opencv-python==4.7.0.68
Pillow==9.4.0
numpy==1.24.1